refactor : 카드 선택 컴포넌트 분리
