+++ presentation/src/main/java/com/nexters/boolti/presentation/screen/gift/GiftScreen.kt
@@ -315,108 +315,6 @@ fun GiftScreen(
 }
 
 @Composable
-private fun CardSelection(
-    message: String,
-    onMessageChanged: (String) -> Unit,
-    images: List<ImagePair>,
-    selectedImage: ImagePair?,
-    onImageSelected: (ImagePair) -> Unit,
-) {
-    Column(
-        modifier = Modifier.padding(top = 24.dp, bottom = 48.dp),
-    ) {
-        Column(
-            modifier = Modifier
-                .padding(horizontal = 32.dp)
-                .clip(RoundedCornerShape(8.dp))
-                .background(
-                    brush = Brush.verticalGradient(
-                        colors = listOf(
-                            Color(0xFFFF5A14),
-                            Color(0xFFFFA883),
-                        )
-                    )
-                )
-                .border(
-                    width = 1.dp,
-                    color = Color(0xFFFFA883),
-                    shape = RoundedCornerShape(8.dp)
-                )
-                .padding(horizontal = 20.dp, vertical = 32.dp)
-                .fillMaxWidth(),
-            horizontalAlignment = Alignment.CenterHorizontally,
-        ) {
-            val maximumLength = 40
-            val messageLengthUnit = stringResource(id = R.string.gift_message_length_unit)
-
-            BasicTextField(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .height(80.dp),
-                value = message,
-                onValueChange = onMessageChanged,
-                textStyle = MaterialTheme.typography.titleLarge.copy(
-                    color = Color.White,
-                    textAlign = TextAlign.Center
-                ),
-                cursorBrush = SolidColor(Color.White)
-            )
-            Text(
-                modifier = Modifier.padding(top = 12.dp),
-                text = "${message.length}/${maximumLength}${messageLengthUnit}",
-                style = MaterialTheme.typography.labelMedium.copy(color = Grey10),
-            )
-
-            AsyncImage(
-                model = selectedImage?.originImage,
-                contentDescription = stringResource(id = R.string.gift_selected_image),
-                modifier = Modifier
-                    .padding(top = 28.dp)
-                    .fillMaxWidth()
-                    .aspectRatio(3 / 2f)
-                    .background(Color.White),
-                contentScale = ContentScale.Crop,
-            )
-        }
-
-        Row(
-            modifier = Modifier
-                .padding(top = 44.dp)
-                .fillMaxWidth(),
-        ) {
-            LazyRow(
-                contentPadding = PaddingValues(start = 32.dp),
-                horizontalArrangement = Arrangement.spacedBy(12.dp),
-            ) {
-                items(images) { image ->
-                    val modifier = if (image == selectedImage) {
-                        Modifier.border(
-                            width = 1.dp,
-                            color = MaterialTheme.colorScheme.primary,
-                            shape = RoundedCornerShape(4.dp)
-                        )
-                    } else {
-                        Modifier
-                    }
-
-                    AsyncImage(
-                        model = image.thumbnailImage,
-                        contentDescription = stringResource(id = R.string.gift_image),
-                        modifier = modifier
-                            .size(52.dp)
-                            .clip(RoundedCornerShape(4.dp))
-                            .clickable {
-                                onImageSelected(image)
-                            },
-                        contentScale = ContentScale.Crop,
-                    )
-                }
-            }
-        }
-    }
-}
-
-@Composable
 private fun ContactInfo(
     sectionName: String,
     name: String,
@@ -461,25 +359,5 @@ private fun ContactInfo(
                 color = Grey40,
             )
         }
-    }
-}
-
-@Preview
-@Composable
-private fun CardSelectionPreview() {
-    BooltiTheme {
-        CardSelection(
-            message = "공연에 초대합니다.",
-            onMessageChanged = {},
-            images = (1..10).map {
-                ImagePair(
-                    it.toString(),
-                    "https://picsum.photos/200",
-                    "https://picsum.photos/200"
-                )
-            },
-            selectedImage = ImagePair("", "", ""),
-            onImageSelected = {}
-        )
     }
 }
